+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.12.0 - Breaking Change
+* Removed `moduleQualifier` from `FlutterDDIModuleDefine`.
+* `routeQualifier` is now the module qualifier API in `FlutterDDIModuleDefine`.
+
 ## 0.11.1
 * Fixed behavior when modules are destroyed.
 * Fixed a possible memory leak when using interceptors.